Allow app to pass in a nonce value to use in the hashing of config files.
This allows the config file in test to be differentiated from the config
file in prod by passing the TARGET_ENV in as the nonce value.  The two
may share the same cookie but when you move back and forth you'll get your
cookie reset to the defaults for that environment.