Update version of vows
npm test was failing with an error inside vows (part of the Behavioral
Driven Design testing framework). I noticed that vows was rather
old and that intervening releases of it mentioned updating for newer
versions of node.  Upgrading vows to this version fixed the tests.
This is the latest released version of vows. Did not use a newer
release that is only in release candidate state.

Closes #12